move lineIntersect -> Lib.geometry2d
